requirements.txt
Flask
requests
streamlit
pandas
psycopg2-binary
